feat(misc): Add docs and workflows for using Conventional commits (#602)
To enable automatic change log generation and having a more uniform
of recording the work done we are introducing the use of conventional
commits. There are docs on how to use this in `CONTRIBUTING.md` and we
added CI workflows to make sure we conform to the standard. PR titles and
descriptions should conform to the standard.