feat(output): Points with no footprints shown as triangles (#604)
The footprints shapefile that was created as output, did not include any
of the buildings that did not have a building footprint assigned to
them. Now, the fiat_toolbox method was updated to make chematized
footprints with standard shapes and diameters. The default shape is a
triangle with 10 m diameter. This can be changed through the site
configuration using the
`fiat.no_footprints.shape` and `fiat.no_footprints.diameter` attributes.
This connects to the fiat_toolbox update
https://github.com/Deltares/fiat_toolbox/pull/62