feat(object-model): refactor site.toml into several files (#630)
the original format of the `site.toml` is now divided into (currently) 4
files:
- `site.toml` is basically an index, saying where to find the other
files. if the files mentioned in there are relative they should be
considered relative to the location of the `site.toml` not to the
execution of the program
- `gui.toml` 
- `sfincs.toml`
- `fiat.toml` 

Hopefully the contents of these files are relatively self explanatory.