refactor(adapters): FIAT adapter refactor (#654)
This is a refactor of the FiatAdapter and DirectImpacts classes,
following the refactor of the Hazard class. More specifically:

- DirectImpacts class is now named Impacts, and only functions as the
Impacts workflow manager.
- FiatAdapter now inlcudes all the methods for preprocessing, running
and postprocessing FIAT simulations.
- Right now the pre-process, run and post-process methods are connected
with FA objects and paths, but the individual methods in the adapter
used are generalized and FA ignorant.
- A FiatColumns class was created in the FIAT adapter which has all the
reference to the exposure columns names. This will make it flexible to
update for different FIAT versions