refactor(adapters): Use pydantic model for defining impact columns formatting (enables use of latest FIAT version (#694)
With this update the latest changes in fiat_toolbox are used. Now a
FiatColumns class defines the expected FIAT columns and their string
formatting. This way, one formatting is tighted to the FIAT output based
on the FIAT version used, and one is tighted to the FloodAdapt output
which can be defined in the FiatAdapter.