feat(adapters): Allow for water depth and area method to be used in FIAT (#705)
This commit makes having building footprints not mandatory and allows
for using FIAT models with polygons which can use the area method for
the calculations, and the waterdepth map as hazard.

Moreover, in case a FIAT model has only point building data and no
footprints are available, the default shapes (now this would be
triangles) will be created.