feat(misc): make floodadapt platform independent (#738)
## Issue addressed
Fixes #642 

## Explanation
Instead of hard coding the paths to the executables as part of the
`system_folder`, just get the paths as arguments in the Settings class.
This will allow users on any platform to compile / prepare their model
and use that path as the bin path.

## Checklist
- [ ] Updated tests or added new tests
- [ ] Branch is up to date with `main`
- [ ] Updated documentation if needed

## Additional Notes (optional)
To be included in v0.3.3