refactor(dbs-controller): Take out edit methods from API and add overwrite option in save method (#768)
## Issue addressed
The edit method was essentially a save with overwrite, which was now
added as an option in the save method, in order to completely take out
all the edit methods for each object.

## Explanation
Explain how you addressed the bug/feature request, what choices you made
and why.

## Checklist
- [x] Updated tests or added new tests
- [x] Branch is up to date with `main`
- [x] Updated documentation if needed

## Additional Notes (optional)
Add any additional notes or information that may be helpful.