test(adapter): improve `add_forcing` tests to test for actual values (#771)
## Issue addressed
Fixes #<issue number>

## Explanation
instead of testing for no errors, we now test for actual values in the
input files
Also improve tests for `add_measure()` by getting file from the config
in the SfincsModel, and testing for the contents of that file


## Checklist
- [ ] Updated tests or added new tests
- [ ] Branch is up to date with `main`
- [ ] Updated documentation if needed

## Additional Notes (optional)
The generated waterlevels from SFINCS are not tested here. 
That is something we should probably do as well, but we are at least
sure the inputs to SFINCS are as expected.