fix(dbs-controller): Not mandatory configs for return periods and SLR scenarios do not break the initialization (#798)
## Issue addressed
Makes sure that FloodAdapt still runs when non-mandatory inputs are not
provided, for return periods and SLR scenarios.

## Explanation
Default values for risk (which has the return periods) has been udpated
from None to the default risk Model, and the method to get the slr
scenarios is only run if the file is provided.

## Checklist
- [x] Updated tests or added new tests
- [x] Branch is up to date with `main`
- [x] Updated documentation if needed

## Additional Notes (optional)
Add any additional notes or information that may be helpful.