refactor(adapter): Updated initializations and FIAT model loadings in the impact adapters to be faster (#805)
## Issue addressed
The way that the impact adapter is initialized (by loading the FIAT
model each time) is making a lot of methods really slow, while the
loading is not needed every time.
This is obvious especially in the Benefits tab of the GUI, which takes a
couple of seconds only for just clicking on the tab and selecting a
benefit.

## Explanation
The FiatAdapter was updated to now have a model property instead of
loading the FIAT model during init. The workflows were updated as well
to use properties and avoid loading in init. The Impact class now has a
models property, and the BenefitRunner has now a scenarios property.

## Checklist
- [x] Updated tests or added new tests
- [x] Branch is up to date with `main`
- [x] Updated documentation if needed

## Additional Notes (optional)
Add any additional notes or information that may be helpful.