refactor(dbs-controller): remove unused api and restructure code (#797)
## Issue addressed
Adresses #329 

## Explanation
Moved from `Database`:
- `interp_slr()` & `plot_slr_scenarios()` to `Slr_ScenariosModel`
- `run_benefit()` & `create_benefit_scenarios()` to `BenefitRunner`

Removed unused functions from `FloodAdapt` api:
- `write_cyc()` & `write_to_csv()`

Updated functions that used to try except -> return a bool to just raise
the error.
- `get_static_map()` & `save_scenario()`

- Added a `exceptions.py` containing `DatabaseError` class for use in
the database so its clear to see that we are raising an error or it is a
bug / some dependency might have raised it.

## Checklist
- [ ] Updated tests or added new tests
- [ ] Branch is up to date with `main`
- [ ] Updated documentation if needed

## Additional Notes (optional)
Add any additional notes or information that may be helpful.

---------

Co-authored-by: Panos Athanasiou <panos.athanasiou@deltares.nl>