refactor(dbs-builder): Update how dbs builder gets default water level references (#833)
## Issue addressed
The default references in the config is a having the "MSL" set to 0. But
we need to always provide a reference for the overland SFINCS model. If
this is not "MSL", the user needs to know that they have to add the
datum in the references list.

## Explanation
Now, the user only needs to provide a name for the reference of the
SFINCS model. Then if the user does not provide manualy a water level
reference obejct, the SFINCS ref name will be used as the main reference
in the water level references object, and assumed=0.

Then the standard updates if a NOAA station is used, will take place

## Checklist
- [ ] Updated tests or added new tests
- [ ] Branch is up to date with `main`
- [ ] Updated documentation if needed

## Additional Notes (optional)
Add any additional notes or information that may be helpful.