fix(dbs-controller): Remove caching of models and check for scenario run instead of hazard run (#854)
## Issue addressed
During one of the last iterations, the models in the dbs_static class
were cached. This created issues during initializations of the database
class.

This was done to speed up some calls in the gui. This was now acchieved
by using the scenario.has_run method instead of the
scenario.hazard_run_check method which is a lot slower.

If the scenario is run, by definition the hazard has been run.

## Explanation
Explain how you addressed the bug/feature request, what choices you made
and why.

## Checklist
- [ ] Updated tests or added new tests
- [ ] Branch is up to date with `main`
- [ ] Updated documentation if needed

## Additional Notes (optional)
Add any additional notes or information that may be helpful.