fix(adapter): Do not crash when tide gauge data for time series is not available (#853)
## Issue addressed
Before it was checked if the return dataframe of the tide gauge
timeseries was None while it is empty when no data is available. Now
both check are in place.

## Explanation
Explain how you addressed the bug/feature request, what choices you made
and why.

## Checklist
- [ ] Updated tests or added new tests
- [ ] Branch is up to date with `main`
- [ ] Updated documentation if needed

## Additional Notes (optional)
Add any additional notes or information that may be helpful.

Co-authored-by: LuukBlom <153174893+LuukBlom@users.noreply.github.com>