feat(object-model): Allow for choosing different spatial infometrics to show (#862)
## Issue addressed
Connects to
https://github.com/Deltares-research/FloodAdapt-GUI/issues/625

## Explanation
A new `MetricLayer` class has been created which allows for filters to
be applied in order to find the metrics that fall within that type of
layer. So you can define a building count layer, a people count or road
length based on strings that are in specific attributes.

## Checklist
- [x] Updated tests or added new tests
- [x] Branch is up to date with `main`
- [ ] Updated documentation if needed

## Additional Notes (optional)
Add any additional notes or information that may be helpful.