fix(misc): updated installation guide and readme and made env creation less strict (#878)
## Issue addressed
fixes first impression of local install

## Explanation
this PR fixes the problems with having to specify your pixi env all the
time.
I got it working by separating the groups of dependencies into features.
these features can then be combined in the environments.
 
all features except build are in the default env, including the editable
install,
for the build env, we include the core and the build dependencies, and a
non-editable install.
 
since there is now 0 ambiguity as to where each task can be run, we dont
have to specify it anymore.
 
for example, pixi run render-docs is part of the docs feature, which is
only included in the default env, so pixi uses that)
alternatively, pixi run build-db-builder is a task for the build
feature, which is only included in the build env, so that works also)
## Checklist
- [ ] Updated tests or added new tests
- [ ] Branch is up to date with `main`
- [ ] Updated documentation if needed

## Additional Notes (optional)
Add any additional notes or information that may be helpful.